ci: compile tests for every published target
Move Unix-only nft verification fixtures behind the Unix build boundary while retaining portable rule validation across targets. Keep emitter health coverage on Windows through the platform stub, then add both Windows architectures to the cross-target vet matrix so test-only compile failures block CI.
